store a drawing as a single file foo.tdraw (is a tar file with foo.draw, foo.map and foo.png inside)
When TWikiDraw is invoked, it does a GET request for the .draw.
Then the drawing is edited and when saving, TWD POSTs 3 files:
.draw, .png and .map

Moin now creates the .tdraw tar file when it gets the .draw POST.
After that it appends the .png and the .map to the archive.

There is a new do=box sub-action of AttachFile that can extract
and send a member of a tar file.