search: improve search result ordering (details below)
Title search orders by pagename. Now we also consider the attachment name,
so that a page hit comes first and is followed by hits for attachments of
that page.

Full text search is ordering by hit weight and pagename. Now we also consider
the attachment name there, so that if weight and pagename is same, the page
hit will come first and is followed by hits for attachments of that page.
This might be a rather rare case, because usually weight is not the same,
but we do it for consistency with title search.