add Expires header with now + 10 seconds for AttachFile do=get and do=box
User's browser shall usually request the wiki content as it might change fast.
It'll either get a 200 OK response with the content, or (if content was not
modified) a 304 NOT MODIFIED response.

Without the Expires header, browser does not request the content (I guess it
uses some builtin heuristics then to determine when to do a request, and
that that might be much too slow for fast changing wikis).