text editor: if edit_rows==0 (user profile or config), dynamically size the text editor height
Accepts anything from 0..999 for edit_rows userprefs now.

Dynamic height is 1.5 * current height. Use Preview if this is not enough.

This is done for devices like the iPhone, when user can't operate double sliders.
It might be comfortable even on other devices to not have to deal with double sliders.