on a save error, show the editor again and highlight the error as such
Previously, the (unsaved) page would be shown again, with a small
message with class="info" at the top. It is easy to miss the save error
like this, especially since you get the page again instead of the
editor. Also, browsers might not always preserve your editor contents
after pressing the back button.

With this patch, the error is clearly marked using the "error" message
class. Also, the editor is shown again (just as if the user clicked
preview), so you can fix the error right away.